Gen 5 personal tables

+++ src/personal_info/traits/personal_encounter.rs
@@ -1,5 +1,5 @@
 pub trait PersonalEncounter {
-    fn base_exp(&self) -> u8;
+    fn base_exp(&self) -> u16;
     fn hatch_cycles(&self) -> u8;
     fn catch_rate(&self) -> u8;
     fn base_friendship(&self) -> u8;
