+++ src/personal_info/traits/personal_misc.rs
@@ -1,6 +1,6 @@
 pub trait PersonalMisc {
     fn evo_stage(&self) -> u8;
     fn color(&self) -> u8;
-    fn height(&self) -> u8;
-    fn weight(&self) -> u8;
+    fn height(&self) -> u16;
+    fn weight(&self) -> u16;
 }
